requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
pandas
numpy
yfinance
plotly
python-multipart
pydantic
